+++ R/adjust_pvalue.R
@@ -2,7 +2,7 @@
 NULL
 #' Adjust P-values for Multiple Comparisons
 #' @description A pipe-friendly function to add an adjusted p-value column into
-#'   a data frame.
+#'   a data frame. Supports grouped data.
 #' @param data a data frame containing a p-value column
 #' @param p.col column name containing p-values
 #' @param output.col the output column name to hold the adjusted p-values
@@ -20,17 +20,29 @@ NULL
 #'
 #' @rdname adjust_pvalue
 #' @export
-adjust_pvalue <- function(data, p.col, output.col, method = "holm"){
+adjust_pvalue <- function(data, p.col = NULL, output.col = NULL, method = "holm"){
+
+  if (is_grouped_df(data)) {
+    res <- data %>%
+      doo(adjust_pvalue, p.col, output.col, method = method)
+    return(res)
+  }
   .attributes <- get_test_attributes(data)
+  if(!is.null(.attributes$args)){
+    .attributes$args$p.adjust.method = method
+  }
   p.adjust <- stats::p.adjust
   p.adjust.method <-  method
 
   # Guess p-value columns if missing
-  if(missing(p.col))
-    p.col <- .guess_pvalue_column(data)
-  if(missing(output.col))
+  if(is.null(p.col))
+    p.col <- data %>% p_detect("p")
+  if(is.null(p.col))
+    return(data)
+  else if(!(p.col %in% colnames(data)))
+    stop("The column ", p.col, "does not exist in the data")
+  if(is.null(output.col))
     output.col <- paste0(p.col, ".adj")
-
   # Adjust p-value
   data %>%
     mutate(