+++ R/tukey_hsd.R
@@ -7,7 +7,8 @@ NULL
 #'
 #'
 #'@description Provides a pipe-friendly framework to performs Tukey post-hoc
-#'  tests. Wrapper around the function \code{\link[stats]{TukeyHSD}()}.
+#'  tests. Wrapper around the function \code{\link[stats]{TukeyHSD}()}. It is
+#'  essentially a t-test that corrects for multiple testing.
 #'
 #'  Can handle different inputs formats: aov, lm, formula.
 #'@param x an object of class \code{aov}, \code{lm} or \code{data.frame}
@@ -59,16 +60,18 @@ tukey_hsd <- function(x, ...){
 #' @describeIn tukey_hsd performs tukey post-hoc test from \code{aov()} results.
 tukey_hsd.default <- function(x, ...)
 {
-  comparison <- comparison2 <- adj.p.value <- p.adj <-
+  comparison <- adj.p.value <- p.adj <-
     term <- group1 <- group2 <- NULL
   res <- TukeyHSD(x, ...) %>%
     broom::tidy() %>%
-    mutate(comparison2 = comparison) %>%
-    separate(comparison2, into= c("group2", "group1"), sep = "-") %>%
+    separate(comparison, into= c("group2", "group1"), sep = "-") %>%
     rename(p.adj = adj.p.value) %>%
     mutate(p.adj = signif(p.adj, 3)) %>%
-    select(term, group1, group2, everything())
-  res
+    select(term, group1, group2, everything()) %>%
+    add_significance("p.adj")
+  res %>%
+    set_attrs(args = list(x = x, p.adjust.method = "Tukey", method = "tukey_hsd")) %>%
+    add_class(c("rstatix_test", "tukey_hsd"))
 }
 
 #' @export
@@ -85,7 +88,10 @@ tukey_hsd.lm <- function(x, ...)
 #'   \code{\link[stats]{aov}()}
 #' @export
 tukey_hsd.data.frame <- function(x, formula, ...){
-  args <- list(data = x, formula = formula, method = "tukey_hsd")
+  args <- list(
+    data = x, formula = formula, method = "tukey_hsd",
+    p.adjust.method = "Tukey"
+    )
   if(is_grouped_df(x)){
     results <- x %>%
       doo(tukey_hsd.data.frame, formula, ...) %>%