+++ R/wilcox_test.R
@@ -1,4 +1,4 @@
-#' @include utilities.R utilities_mean_test.R
+#' @include utilities.R utilities_two_sample_test.R
 #' @importFrom stats wilcox.test
 NULL
 #'Wilcoxon Tests
@@ -17,8 +17,10 @@ NULL
 #'  specified, for a given grouping variable, each of the group levels will be
 #'  compared to the reference group (i.e. control group).
 #'
-#'  If \code{ref.group = "all"}, pairwise two sample Wilcoxon tests are performed for comparing each grouping
-#'  variable levels against all (i.e. basemean).
+#'  If \code{ref.group = "all"}, pairwise two sample tests are
+#'  performed for comparing each grouping variable levels against all (i.e.
+#'  basemean).
+#'@param mu a number specifying an optional parameter used to form the null hypothesis.
 #'@param comparisons A list of length-2 vectors specifying the groups of
 #'  interest to be compared. For example to compare groups "A" vs "B" and "B" vs
 #'  "C", the argument is as follow: \code{comparisons = list(c("A", "B"), c("B",
@@ -28,12 +30,13 @@ NULL
 #'  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". If you don't
 #'  want to adjust the p value (not recommended), use p.adjust.method = "none".
 #'
-#' @param detailed logical value. Default is FALSE. If TRUE, a detailed result is shown.
+#'@param detailed logical value. Default is FALSE. If TRUE, a detailed result is
+#'  shown.
 #'@param ... other arguments to be passed to the function
 #'  \code{\link[stats]{wilcox.test}}.
 #'
-#'@details - \code{pairwise_wilcox_test()} applies the standard two sample Wilcoxon test to
-#'  all possible pairs of groups. This method calls the
+#'@details - \code{pairwise_wilcox_test()} applies the standard two sample
+#'  Wilcoxon test to all possible pairs of groups. This method calls the
 #'  \code{\link[stats]{wilcox.test}()}, so extra arguments are accepted.
 #'
 #'
@@ -44,16 +47,17 @@ NULL
 #'  - For a grouped data, if pairwise test is performed, then the p-values are
 #'  adjusted for each group level independently.
 #'
-#'@return return a data frame with some of the following columns: \itemize{ \item
-#'  \code{.y.}: the y variable used in the test. \item \code{group1,group2}: the
-#'  compared groups in the pairwise tests. \item \code{statistic}: Test
-#'  statistic used to compute the p-value. \item \code{p}: p-value. \item
-#'  \code{p.adj}: the adjusted p-value. \item \code{method}: the statistical
-#'  test used to compare groups. \item \code{p.signif, p.adj.signif}: the
-#'  significance level of p-values and adjusted p-values, respectively. }
+#'@return return a data frame with some of the following columns: \itemize{
+#'  \item \code{.y.}: the y variable used in the test. \item
+#'  \code{group1,group2}: the compared groups in the pairwise tests. \item
+#'  \code{n,n1,n2}: Sample counts. \item \code{statistic}: Test statistic used
+#'  to compute the p-value. \item \code{p}: p-value. \item \code{p.adj}: the
+#'  adjusted p-value. \item \code{method}: the statistical test used to compare
+#'  groups. \item \code{p.signif, p.adj.signif}: the significance level of
+#'  p-values and adjusted p-values, respectively. }
 #'
-#'  The \strong{returned object has an attribute called args}, which is a list holding
-#'  the test arguments.
+#'  The \strong{returned object has an attribute called args}, which is a list
+#'  holding the test arguments.
 #' @examples
 #' # Load data
 #' #:::::::::::::::::::::::::::::::::::::::
@@ -96,7 +100,7 @@ NULL
 #' #::::::::::::::::::::::::::::::::::::::::
 #' df %>% wilcox_test(len ~ dose, ref.group = "all")
 #'
-#' @describeIn wilcox_test Wilcoxon test
+#'@describeIn wilcox_test Wilcoxon test
 #'@export
 wilcox_test <- function(
   data, formula, comparisons = NULL, ref.group = NULL,
@@ -105,75 +109,29 @@ wilcox_test <- function(
   mu = 0, conf.level = 0.95, detailed = FALSE
 )
 {
-  args <- as.list(environment()) %>%
+  env <- as.list(environment())
+  args <- env %>%
     .add_item(method = "wilcox_test")
+  params <- env %>%
+    remove_null_items() %>%
+    add_item(method = "wilcox.test")
 
   outcome <- get_formula_left_hand_side(formula)
   group <- get_formula_right_hand_side(formula)
   number.of.groups <- guess_number_of_groups(data, group)
-
-  # Case of one sample test
-  if(number.of.groups == 1){
-    res <- one_sample_wilcox_test(
-      data = data, formula = formula, exact = exact,
-      alternative = alternative, mu = mu,
-      conf.level = conf.level, detailed = detailed
-    )
+  if(number.of.groups > 2 & !is.null(ref.group)){
+    if(ref.group %in% c("all", ".all.")){
+      params$data <- create_data_with_all_ref_group(data, outcome, group)
+      params$ref.group <- "all"
+    }
   }
-  # Case of two independents or paired groups
-  else if (number.of.groups == 2) {
-    res <- two_sample_wilcox_test(
-      data = data, formula = formula,
-      paired = paired, exact = exact,
-      alternative = alternative,
-      conf.level = conf.level, ref.group = ref.group,
-      detailed = detailed
-    )
-  }
-  # Pairwise comparisons
-  else if(number.of.groups > 2){
-
-    if(is.null(ref.group))
-      res <- pairwise_wilcox_test(
-        data = data, formula = formula,
-        comparisons = comparisons,
-        p.adjust.method = p.adjust.method,
-        paired = paired, exact = exact,
-        alternative = alternative, conf.level = conf.level,
-        detailed = detailed
-      )
-    else if(ref.group %in% c("all", ".all."))
-      res <- one_vs_all_wilcox_test(
-        data = data, formula = formula,
-        p.adjust.method = p.adjust.method,
-        exact = exact, alternative = alternative,
-        conf.level = conf.level, detailed = detailed
-      )
-    else
-      res <- pairwise_wilcox_test(
-        data = data, formula = formula,
-        comparisons = comparisons, ref.group = ref.group,
-        p.adjust.method = p.adjust.method,
-        paired = paired, exact = exact,
-        alternative = alternative, conf.level = conf.level,
-        detailed = detailed
-      )
-  }
-  res %>%
+  test.func <- two_sample_test
+  if(number.of.groups > 2) test.func <- pairwise_two_sample_test
+  do.call(test.func, params) %>%
     set_attrs(args = args) %>%
     add_class(c("rstatix_test", "wilcox_test"))
 }
 
-
-
-one_sample_wilcox_test <- function(data, formula, mu = 0, exact = FALSE, ...){
-  mean_test(data, formula, method = "wilcox.test", mu = mu, ...)
-}
-
-two_sample_wilcox_test <- function(data, formula, paired = FALSE,  ...)
-{
-  mean_test(data, formula, method = "wilcox.test", paired = paired, ...)
-}
 
 
 #'@describeIn wilcox_test performs pairwise two sample Wilcoxon test.
@@ -184,7 +142,7 @@ pairwise_wilcox_test <- function(
   {
   args <- as.list(environment()) %>%
     .add_item(method = "wilcox_test")
-  res <- mean_test_pairwise(
+  res <- pairwise_two_sample_test(
     data, formula, method = "wilcox.test",
     comparisons = comparisons, ref.group = ref.group,
     p.adjust.method = p.adjust.method, detailed = detailed, ...
@@ -192,16 +150,4 @@ pairwise_wilcox_test <- function(
   res %>%
     set_attrs(args = args) %>%
     add_class(c("rstatix_test", "wilcox_test"))
-}
-
-
-# wilcox_test performs pairwise two sample Wilcoxon test comparing each grouping
-#  variable levels against all (i.e. basemean)
-one_vs_all_wilcox_test <- function(data, formula, p.adjust.method = "holm", ...)
-{
-
-  mean_test_one_vs_all (
-    data, formula, method = "wilcox.test",
-    p.adjust.method = p.adjust.method, ...
-  )
 }